simplify repro testcase.
